Float and Bool criteria
