+++ Assets/Scripts/Incidents/IIncidentCriteria.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b83a7b38f0a6ec744bc1d663d2992692
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
